Create RightTriangleChecker.java
Right Triangle Checker

Write a program to allow the user to enter three integers. You must use do-while or while loops to enforce that these integers are in ascending order, though duplicate numbers are allowed.

Tell the user whether or not these integers would represent the sides of a right triangle.