Create BabyNim.java
Write a program that starts with three "piles" of 3 counters each. Let the player choose piles and remove counters until all the piles are empty.

Start by placing counters (coins or toothpicks or something) into 3 piles.
The player picks a pile, then removes one or more counters from that pile. (It's okay to take the whole pile.)
The player picks a new pile, then removes one or more counters from that pile. (It's okay to pick the same pile as before.)
Once all piles are empty, the game stops.
You do not need to check for errors like a wrong pile name, or if someone tries to take more counters from the pile than the pile has.